+++ integration_tests/test_parallelism/test_parallelism.py
@@ -4,11 +4,12 @@ import glob
 
 from send2trash import send2trash
 
-from integration_test import IntegrationTest
-from util import (copy_file, copy_dir, import_tests, execute_tests)
+from integration_tests.integration_test import IntegrationTest
+from integration_tests.util import (copy_file, copy_dir, import_tests,
+                                    execute_tests)
 
 
-class TestParallelism(IntegrationTest):
+class Parallelism(IntegrationTest):
     """Test Parallel deployment scenarios."""
 
     base_dir = os.path.abspath(os.path.dirname(__file__))
@@ -16,10 +17,6 @@ class TestParallelism(IntegrationTest):
     tests_dir = os.path.join(base_dir, 'tests')
 
     parallelism_test_dir = os.path.join(base_dir, 'parallelism_test')
-
-    def __init__(self, logger):
-        """Initialize."""
-        IntegrationTest.__init__(self, logger)
 
     def copy_fixture(self, name='two-regions-app.cfn'):
         """Copy fixture files for test."""
@@ -36,7 +33,9 @@ class TestParallelism(IntegrationTest):
     def run(self):
         """Find all tests and run them."""
         import_tests(self.logger, self.tests_dir, 'test_*')
-        tests = [test(self.logger) for test in TestParallelism.__subclasses__()]
+        tests = [test(self.logger) for test in Parallelism.__subclasses__()]
+        if not tests:
+            raise Exception('No tests were found.')
         self.logger.debug('FOUND TESTS: %s', tests)
         self.set_environment('dev')
         err_count = execute_tests(tests, self.logger)