+++ integration_tests/test_parallelism/tests/test_default_to_series.py
@@ -1,24 +1,19 @@
 """Testing that when CI is not set it still will launch the regions in series"""
-
-from test_parallelism.test_parallelism import TestParallelism
 from runway.util import change_dir
-from util import run_command
+
+from integration_tests.test_parallelism.test_parallelism import Parallelism
+from integration_tests.util import run_command
 
 
-class TestDefaultToSeries(TestParallelism):
+class TestDefaultToSeries(Parallelism):
     """Test deploying two regions in series."""
 
     TEST_NAME = __name__
 
-    def __init__(self, logger):
-        """Init class."""
-        TestParallelism.__init__(self, logger)
-        self.logger = logger
-
     def deploy(self):
         """Deploy provider."""
-        self.copy_runway('default-to-series')
         self.copy_fixture('sampleapp.cfn')
+        self.copy_runway('default-to-series')
         with change_dir(self.parallelism_test_dir):
             return run_command(['runway', 'deploy'])