+++ integration_tests/test_terraform/test_terraform.py
@@ -5,8 +5,10 @@ import glob
 import platform
 from send2trash import send2trash
 from runway.util import change_dir
-from integration_test import IntegrationTest
-from util import (copy_file, import_tests, execute_tests, run_command)
+
+from integration_tests.integration_test import IntegrationTest
+from integration_tests.util import (copy_file, import_tests,
+                                    execute_tests, run_command)
 
 
 class Terraform(IntegrationTest):
@@ -18,9 +20,6 @@ class Terraform(IntegrationTest):
 
     tf_test_dir = os.path.join(base_dir, 'terraform_test.tf')
     tf_state_dir = os.path.join(base_dir, 'tf_state.cfn')
-
-    def __init__(self, logger):
-        IntegrationTest.__init__(self, logger)
 
     def copy_runway(self, template):
         """Copy runway template to proper directory."""
@@ -80,10 +79,16 @@ class Terraform(IntegrationTest):
         """Find all Terraform tests and run them."""
         import_tests(self.logger, self.tests_dir, 'test_*')
         tests = [test(self.logger) for test in Terraform.__subclasses__()]
+        if not tests:
+            raise Exception('No tests were found.')
         self.logger.debug('FOUND TESTS: %s', tests)
         self.set_environment('dev')
-        return execute_tests(tests, self.logger)
+        self.set_env_var('CI', '1')
+        err_count = execute_tests(tests, self.logger)
+        assert err_count == 0  # assert that all subtests were successful
+        return err_count
 
     def teardown(self):
         """Teardown resources create during init."""
+        self.unset_env_var('CI')
         self.clean()