+++ integration_tests/test_terraform/tests/test_local_to_s3_backend.py
@@ -1,18 +1,13 @@
 """Test changing backends between local and S3."""
 from runway.util import change_dir
-from test_terraform.test_terraform import Terraform
-from util import run_command
+from integration_tests.test_terraform.test_terraform import Terraform
+from integration_tests.util import run_command
 
 
 class LocalToS3Backend(Terraform):
     """Test changing between local and S3 backends."""
 
     TEST_NAME = __name__
-
-    def __init__(self, logger):
-        """Init class."""
-        Terraform.__init__(self, logger)
-        self.logger = logger
 
     def deploy_backend(self, backend):
         """Deploy provider."""