+++ integration_tests/test_terraform/tests/test_provider.py
@@ -1,18 +1,13 @@
 """Test changing provider versions."""
 from runway.util import change_dir
-from test_terraform.test_terraform import Terraform
-from util import run_command
+from integration_tests.test_terraform.test_terraform import Terraform
+from integration_tests.util import run_command
 
 
 class ProviderTest(Terraform):
     """Test changing between provider versions."""
 
     TEST_NAME = __name__
-
-    def __init__(self, logger):
-        """Init class."""
-        Terraform.__init__(self, logger)
-        self.logger = logger
 
     def deploy_provider(self, version):
         """Deploy provider."""