+++ src/runway/templates/sls-tsc/src/helloWorld.ts
@@ -1,0 +1,29 @@
+/**
+ * hello-world API
+ */
+
+/** imports (delete comment after https://github.com/TypeStrong/typedoc/issues/603 resolution) */
+import {
+  APIGatewayProxyEvent,
+  APIGatewayProxyHandler,
+  APIGatewayProxyResult,
+  Context,
+} from "aws-lambda";
+import "source-map-support/register";
+
+/** AWS Lambda entrypoint */
+export let handler: APIGatewayProxyHandler = async (
+  event: APIGatewayProxyEvent,
+  context: Context,
+): Promise<APIGatewayProxyResult> => {
+  return {
+    statusCode: 200,
+    headers: {
+      "Access-Control-Allow-Origin": "*", // Required for CORS support to work
+    },
+    body: JSON.stringify({
+      message: "Go Serverless v1.0! Your function executed successfully!",
+      input: event,
+    }),
+  };
+};